Fix for fmGo
Webviewers don't seem to update unless the hash is unique. This is problematic if you want to repeatedly call a js function with the same parameters. Added a uuid property to the url hash in the Compile URL script. This should make the hash unique. An alternative is to use Get ( CurrentTimeUTCMilliseconds ).