Adds Test Dependencies
This commit adds dependencies to enable automated testing and code coverage reporting to this project.

This commit adds development dependencies of nyc, mocha, chai, and mocha-lcov-reporter.

This commit fixes the npm test script to use the added dependencies and adds an additional script, the coverage script, to enable coverage reporting of this project.

Signed-off-by: Lui de la Parra <Lui@mutesymphony.com>